Percentage info retention changed to eigenvector retention for 2dpca

+++ C++/two_d_pca.h
@@ -26,7 +26,7 @@ class pca2d
 		cv::Mat eigenvectors_X;
 
 	public:
-		void train(std::vector<cv::Mat>,std::vector<int>,double,std::string);
+		void train(std::vector<cv::Mat>,std::vector<int>,int,std::string);
 		void load(std::string);
 		int predict(cv::Mat,double distance_thresh=200000000); //keep one zero less
